remove the ad-hoc stuff
